powerpc/powernv: Setup PCI P2P on resource mapping if needed
This patch implements the dma_map_resource callback on the PCI
controller for powernv. It allows to catch a cross-PHB mmio mapping,
which needs to be setup in hardware by calling opal.

Both the initiator and target PHBs need to be configured, so we look
for which PHB owns the mmio address being mapped.

Signed-off-by: Frederic Barrat <fbarrat@linux.ibm.com>